requirements.txt
Flask
flask-restful
Flask-SQLAlchemy
marshmallow-sqlalchemy
Flask-Migrate
gunicorn
flask-jwt-extended